Week 1 Image Classifier
This is a custom image classifier that you can train on the fly in the browser using TensorFlow.js. We are using transfer learning to create a highly accurate model with minimal training data. We are using a pre-trained model for image classification called MobileNet. All you need to do is show some images, add those images to the corresponsing class and model is trained.